+++ CHANGELOG.rst
@@ -8,7 +8,7 @@ v1.8.3
 
 - Added utilities to download files
 - Get text for InputText element in mobile tests
-
+- Add dataset utilities
 
 v1.8.2
 ------